# backend/src/main/java/com/calbooking/model/EventType.java
package com.calbooking.model;

public class EventType {
    private Long id;
    private String title;
    private String description;
    private int durationMinutes;
    private boolean active;

    public EventType() {}

    public EventType(Long id, String title, String description, int durationMinutes, boolean active) {
        this.id = id;
        this.title = title;
        this.description = description;
        this.durationMinutes = durationMinutes;
        this.active = active;
    }

    public Long getId() { return id; }
    public void setId(Long id) { this.id = id; }

    public String getTitle() { return title; }
    public void setTitle(String title) { this.title = title; }

    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }

    public int getDurationMinutes() { return durationMinutes; }
    public void setDurationMinutes(int durationMinutes) { this.durationMinutes = durationMinutes; }

    public boolean isActive() { return active; }
    public void setActive(boolean active) { this.active = active; }
}